Fix 30+ security, concurrency, and stability issues from code review
CRITICAL fixes:
- WorldManager.restore: atomic replace via .bak backup; recover on move failure
- Path traversal: ModManager.resolveJar, ModpackManager.deleteModpack,
  LaunchProfileBuilder.findVersionJson/findVersionJar/resolveGameDir,
  VersionInstaller.install entry validation
- readAllBytes blocking: 5 sites in JavaRuntimeFinder/OptiFineInstaller/
  ProcessTuner now use daemon drainer thread + waitFor(timeout)
- MusicPlayer race: stop() joins thread, loop checks thread identity,
  FFmpegFrameGrabber exception-safe close
- VideoCallSession ICE SSRF: reject loopback/private/link-local candidates
- UI thread blocking: switchMusicPlaylist persists playlist async
- PluginPackageParser: reject ".." in source/lib/resource paths

HIGH fixes:
- SsrfChecker: add IPv6 ULA (fc00::/7) check
- MicrosoftAuthFlow: interval/expires_in defaults lower-bounded
- CurlFallback: add "--" before url to block option injection
- ModpackManager.downloadModpackFile: SSRF validation on download URL
- CurseForgeClient: handle HTTP 429 + Retry-After header
- ModrinthClient.buildFacets: use Gson JsonArray instead of string concat
- PmclHostServer.handleLaunch: convert to suspend, move network I/O off
  WebSocket coroutine; handleKill no longer clears launchClaimed when
  currentProcess is null (fix TOCTOU double-launch)
- News cover image fetch: Mutex-protected read-modify-write on StateFlow
- AudioCache: SSRF validation + 500MB download size cap
- VideoCallSession.iceAgent: volatile for cross-thread visibility
- PluginPackageBuilder: normalize + startsWith sandbox check on LibraryRef
- Atomic writes: InstanceManager.saveInstanceInfo, ConfigFileManager.writeFile
- ModManager: merge dual volatile fields into single CacheEntry
- ModFile/ModMeta: defensive copy on construction and getter
- FriendStore.addMessage: synchronized for atomic dedup+add+trim
- PluginManager.loadState: synchronized to race with saveState
- RuntimeManager: systemInfo/hardware/os now final
- DownloadManager.speedLimitBytesPerSec: int -> long (overflow fix)
- MirrorManager.type/customBase: volatile visibility
- ProcessTuner.caffeinateProcess: volatile visibility
- ModUpdateChecker: null-safe getGameVersions() in version match
- DownloadQueueManager: generation check before removing runningFuture

MEDIUM fixes:
- UI emoji replaced with Material Icons or plain text
- DownloadPage LazyColumn keys: fallback to index when id is null
- GitHubReleaseSyncChecker: regex-validate githubRepo format
- TokenEncryptor.decrypt: return null on failure (was empty string);
  call sites in AuthService/FriendStore/Preferences/PairingManager updated

+++ core/src/main/java/com/pmcl/core/download/ChunkedDownloader.java
@@ -65,7 +65,7 @@ public final class ChunkedDownloader {
     private volatile OkHttpClient http;
     private final int chunkCount;
     private final ExecutorService pool;
-    private volatile int speedLimitBytesPerSec = 0;
+    private volatile long speedLimitBytesPerSec = 0;
 
     /**
      * @param http       共享 OkHttpClient（必须配 Dispatcher.maxRequestsPerHost ≥ chunkCount）
@@ -87,7 +87,7 @@ public final class ChunkedDownloader {
     }
 
     /** 设置限速（bytes/sec，0=不限） */
-    public void setSpeedLimit(int bytesPerSec) {
+    public void setSpeedLimit(long bytesPerSec) {
         this.speedLimitBytesPerSec = Math.max(0, bytesPerSec);
     }
